Fix ChatLib.say not triggering MessageSent
